Replace tar chain with export-time modifications
The process of untarring the Docker-exported rootfs -- modifying it to
remove .dockerenv and adding /etc/scw-release -- is removed in favor of
making those changes directly in the container from which the rootfs is
exported. This has the double benefit of resolving the issue of
incompatibility between Docker's tar export format and bsd tar, while
also removing the need to be root when running the build process (which
was needed for the untarring to work -- as it created files owned by
users with UID < 1000).