systemd: revise scw-set-hostname to fix errors thrown
`hostnamectl` can return more than one hostname, the pair "transient"
and "static" hostname is not uncommon.

The script then fails because `$old_hostname` is set to two lines (!)
with the first being "n/a". Both get passed to `sed` and result in an
invalid script both ways, because `grep` finds it erroneously due to
its special meaning.
      In that case a `hostnamectl --static` won't return 'localhost',
but is indeed empty.

Just assume it's not in /etc/hosts in those cases and stick to writing
the new hostname.